% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rig.R
\name{rig_example_data}
\alias{rig_example_data}
\title{Example Dataset}
\usage{
rig_example_data()
}
\value{
example dataset that could be analysed with rig_predict
}
\description{
Example Dataset
}
\examples{
rig_example_data()
}
