% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rig.R
\name{rig_genes}
\alias{rig_genes}
\title{Genelist}
\usage{
rig_genes(published_only = FALSE)
}
\arguments{
\item{published_only}{forces return of a single vector containing the originally published rig gene names only (no alts)}
}
\value{
a named list where indexes are gene names used in the original publication, and elements represents a vector of valid gene names for that gene.
}
\description{
Genelist
}
\examples{
rig_genes()
}
