% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rig.R
\name{tailor_gene_names_to_signature}
\alias{tailor_gene_names_to_signature}
\title{Tailor Gene Names from a Signature Gene Set to Your Dataset}
\usage{
tailor_gene_names_to_signature(
  dataset_genes,
  signature_gene_set = rig_genes(),
  require_all_genes_found = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{dataset_genes}{A character vector of gene names from your expression dataset.}

\item{signature_gene_set}{A named list where each element is a character vector of valid gene names for a single gene, and names are the common names for those genes.}

\item{require_all_genes_found}{A boolean indicating whether an error should be thrown if there are genes in the signature gene set that cannot be found in the expression dataset. Defaults to TRUE.}

\item{verbose}{verbose (flag)}
}
\value{
A character vector of gene names that are present in 'dataset_genes' and correspond to the signature_gene_set.
}
\description{
This function takes a list of gene names from an expression dataset and tailors them to match a signature gene set.
The function returns the gene names from the dataset that correspond to the signature gene set.
}
